+++ mlrun/runtimes/utils.py
@@ -11,10 +11,37 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-
-from .base import RunError
+import hashlib
 from sys import stderr
+import pandas as pd
+from io import StringIO
 from ..utils import logger
+from .generators import selector
+from ..utils import get_in
+from ..artifacts import TableArtifact
+
+
+class RunError(Exception):
+    pass
+
+
+def calc_hash(func, tag=''):
+    # remove tag, hash, date from calculation
+    tag = tag or func.metadata.tag
+    status = func.status
+    func.metadata.tag = ''
+    func.metadata.hash = ''
+    func.status = None
+    func.metadata.updated = None
+
+    data = func.to_json().encode()
+    h = hashlib.sha1()
+    h.update(data)
+    hashkey = h.hexdigest()
+    func.metadata.tag = tag
+    func.metadata.hash = hashkey
+    func.status = status
+    return hashkey
 
 
 def log_std(db, runobj, out, err='', skip=False):
@@ -48,3 +75,78 @@ class AsyncLogWriter:
     def flush(self):
         # todo: verify writes are large enough, if not cache and use flush
         pass
+
+
+def add_code_metadata(labels):
+    dirpath = './'
+    try:
+        from git import Repo
+        from git.exc import GitCommandError, InvalidGitRepositoryError
+    except ImportError:
+        return
+
+    try:
+        repo = Repo(dirpath, search_parent_directories=True)
+        remotes = [remote.url for remote in repo.remotes]
+        if len(remotes) > 0:
+            set_if_none(labels, 'repo', remotes[0])
+            set_if_none(labels, 'commit', repo.head.commit.hexsha)
+    except (GitCommandError, InvalidGitRepositoryError):
+        pass
+
+
+def set_if_none(struct, key, value):
+    if not struct.get(key):
+        struct[key] = value
+
+
+def results_to_iter(results, runspec, execution):
+    if not results:
+        logger.error('got an empty results list in to_iter')
+        return
+
+    iter = []
+    failed = 0
+    running = 0
+    for task in results:
+        state = get_in(task, ['status', 'state'])
+        id = get_in(task, ['metadata', 'iteration'])
+        struct = {'param': get_in(task, ['spec', 'parameters'], {}),
+                  'output': get_in(task, ['status', 'results'], {}),
+                  'state': state,
+                  'iter': id,
+                  }
+        if state == 'error':
+            failed += 1
+            err = get_in(task, ['status', 'error'], '')
+            logger.error('error in task  {}:{} - {}'.format(
+                runspec.metadata.uid, id, err))
+        elif state != 'completed':
+            running += 1
+
+        iter.append(struct)
+
+    df = pd.io.json.json_normalize(iter).sort_values('iter')
+    header = df.columns.values.tolist()
+    summary = [header] + df.values.tolist()
+    item, id = selector(results, runspec.spec.selector)
+    task = results[item] if id and results else None
+    execution.log_iteration_results(id, summary, task)
+
+    csv_buffer = StringIO()
+    df.to_csv(
+        csv_buffer, index=False, line_terminator='\n', encoding='utf-8')
+    execution.log_artifact(
+        TableArtifact('iteration_results',
+                      src_path='iteration_results.csv',
+                      body=csv_buffer.getvalue(),
+                      header=header,
+                      viewer='table'))
+    if failed:
+        execution.set_state(
+            error='{} tasks failed, check logs in db for details'.format(
+                failed), commit=False)
+    elif running == 0:
+        execution.set_state('completed', commit=False)
+    execution.commit()
+
